Move away from smart pointer in public API
Smart pointers have been removed from the public facing parts of the API
as the overhead of copying objects seemed wasetfull. This allows
modellers to create model objects on either the stack or heap (with
their own management of memory). The only condition that this requires
is that the model descriptions remain valid for the life of the
application.

Smart pointers are still used within the API objects to replace the use
of new and delete. For example containers of object references will use
smart pointers (specifically a unique pointer) and a raw reference to
smart pointer object can be safely passed with a lifetime of the parent
object (owner of the container).

The following has a great guide on use of smart pointers (or more
specifically shared pointers).

https://msdn.microsoft.com/en-gb/library/hh279669.aspx