Pauls Latest changes
Summary:
Moved all hashing functionaility required by the singleton into a new
modified version of cuRVE. This creates a clean serpation of all the
hashing stuff we have to do from the rest of FLAME GPU.

TODO:
I have broken one of the simulation tests (commented out so that the
rest compiles). We need to discuss this one.

The hashing should be removed fromt he CUDA Agent completlely as it is
now redundant. I will do this.

The mapping and unmapping of variable before a function is called still
needs to be done by calling cuRVE functions. This should be straight
forward.