singleton design pattern changed
The FLAME_API object needs to be accessible by the device so the
singleton design does not work for this object.

curve needs to initialised only once (regardless of how many
CUDAAgentModels exist in the program. curve has therefore been wrapped
in a new singleton class called curveinstane. Each CUDAgentModel will
request an instance but curve initialisation will be done only once for
the program.