compilation now fixed
I think. Commit form my laptop without boost so wont fully compile. This
seems to have fixed the errors in the FLAMEGPU_API class however.
Problem was due to simply having the wrong function declaration
(different form the definition). Should be enough now to test some basic
agent functions with gets and sets of variables.

Only point that seems missing is the function pointer syntax in
FLAMEGPU_API does not match that of the AgentfunctionDescription.