Dev/1 4 (#38)
* sim func not wrkng properly-broken

* added func pointer and unit test

* added func to sim

* Pauls comments and changes which break the code! Wont build

* some comments on how FLAMEGPU_APIclass should work

* broken sim

* broken sim fixed-hvn't tested vars

* Broken-singleton class

* Paul added notes on hashing

* broken singleton class

* broken-singleton class

* commented out part of singleton class

* comments for singleton class

* cleaned up

* broken-cudaagent-step func

* Small changed to allow build success

Some changes to simulation functions to make some functions const where
required.

Some explicit casting of size_t values to avoid compiler warnings.

* Pauls Latest changes

Summary:
Moved all hashing functionaility required by the singleton into a new
modified version of cuRVE. This creates a clean serpation of all the
hashing stuff we have to do from the rest of FLAME GPU.

TODO:
I have broken one of the simulation tests (commented out so that the
rest compiles). We need to discuss this one.

The hashing should be removed fromt he CUDA Agent completlely as it is
now redundant. I will do this.

The mapping and unmapping of variable before a function is called still
needs to be done by calling cuRVE functions. This should be straight
forward.

* Hashing removed from CUDA Agent

This has now been successully removed (and all tests pass). All hashing
is now done by the modified version of curve.

* minor fix

updated cuda runitme import

* curve integration

curve maping and unmapping of variables implemented but untested

* minor bug fixes

* fixed a typo

* singleton design pattern changed

The FLAME_API object needs to be accessible by the device so the
singleton design does not work for this object.

curve needs to initialised only once (regardless of how many
CUDAAgentModels exist in the program. curve has therefore been wrapped
in a new singleton class called curveinstane. Each CUDAgentModel will
request an instance but curve initialisation will be done only once for
the program.

* whoops missed an escaped line

* added runtime module

This now instanciates the FLAME_API class which is causing errors. These
existed before but as the class was not used did not give build errors.

* compilation now fixed

I think. Commit form my laptop without boost so wont fully compile. This
seems to have fixed the errors in the FLAMEGPU_API class however.
Problem was due to simply having the wrong function declaration
(different form the definition). Should be enough now to test some basic
agent functions with gets and sets of variables.

Only point that seems missing is the function pointer syntax in
FLAMEGPU_API does not match that of the AgentfunctionDescription.

* unified the use fo FLAMEGPU function pointers

Now consistent accross runtime api header and the agent function
description class.

Missing: Need to be able to get an agent name from a function
description.

To Complete: Step function which is partially complete but needs the
above. This might require a redesign of the AgentFunctionDescription
object.

* boost test fixed

Update example FLAME GPU functions.

* changes to cudaAgentModel step function
Also made changes to the makefile. The codes compiles but there is an issue in linking. So, it couldn't be build.

* fixed makefile

* fixed linking error
Made a few changes to the makefile and changed the type of CUDAAgentModel file to a CUDA file
Todo : running some test to see if the function_wrapper works

* simulation func testing
At this stage, tests passes when function is called (printing hello only).

* minor changes

* updated vs project

* decl func pointer as device var-in progress

* lots of changes to try and get device function pointers to work

Still broken. Lots of cheating happending. Make sure we remove the
includes of the test functions.

* added device var, still works for one specific func

* Funcptr issue on device-FIXED
FLAMEGPU functions are stored in a header rather than main.cu.This commit fixes the issue with simulating functions on device.

* fixed simulation test

* func pointer get/set variable - fix
With this commit set/getVariable functions in FLAMEGPU_API class works fine. However, it still requires type checking.

* minor changes to boost unit tests

* minor change

* changes to boost tests etc

* new FLAMEGPU function macro fixed

Adds the ability to have simple user defined functions without worrying
about function pointers

* boost unit test compilation with nvcc-fixed

* problem with unit testing for GPU

* minor change

* testing get/set

* added unit test for VS-not tested

* type checking curve func-not done

* fixed but not tidy

d_variables was of type float so pointer arithmetic was wrong. Have
changed to char as pointer arithmetic is not perminted on type void.

* type-checking curve

* curve type checking

* type error checking-not working
Note: we should store pointers to type_info object. These objects returned by typeif have static storage

* broken type checker

* unit test changed

* separate unit tests for gpu and sim

* Type checking not working

Type checking needs some investigation. We can not have type_info on the
device. Must be some other represenation of the type (either a string or
hashed string).

Boost unit tests (or anything referencing boost) shoudl not be in the
cuda files.

* curve data type sizes added. ptr lengths still to do.

* fixed boost test units

* vector lenght checking - not tested

* added type and vector checking
core dump occurs in get or set variable - need investigation

* temp fix on core dump

* Checker for Type and Vector Length fixed

* added multi agent example-Not tested

* small fix on example model

* minor change-setNamespace in progress

* added agent name hashing

* setting namespace

* Agent name hashing. We will not use the curve namespace stuff as it is not cuda stream compatible.

* namespace Fixed

Next step would be to solve #25

* stream not wrkng properly

* adding comments to the test suites

* added comments to pop validation tests

* streams

* minor changes to streams

* Pauls Comments on concurrency

* added db to ignore

* cuda stream wrkng

* commenting tests [WIP]

* fixed gpu unit test

* readme

* commenting on tests+modifying tests [WIP]

* binding agent,func,var names

* Review with Paul regarding testing

* fixed tests on VS

* dev1 RFR

* cleanup

* cleanup

* slight change to VS project filters for new file structure