Suggested fix for segfault in simple_example.
Implicitly calls CUDAMessage::setInitialMessageList() at construction of CUDAMessage.

Messages are not filled on the host, so initialisation should be implicit. Only concern is whether the list size is allocated to match agent size, I didn't check that far.

Comments out CUDAAgentModel::setMessageData(MessageDescription&) as it is now redundant.

It should be removed before this is merged.

Also adds a check/exception to the offending method incase pointer is empty.

Now executes to exit without segfault (on Windows).