Fix for segfault in simple_example. (#119)
Implicitly calls CUDAMessage::setInitialMessageList() at construction of CUDAMessage.
Messages are not filled on the host, so initialisation should be implicit. 
Removes CUDAAgentModel::setMessageData(MessageDescription&) as it is now redundant.
Also adds a check/exception to the offending method incase pointer is empty.
Now executes to exit without segfault (on Windows).