At this stage, a user must specify the message input and output type when adding the function to a layer (or similar).
This is necessary to instantiate the correct agent_function_wrapper

The user must also specify the MsgIn and MsgOut types when delcaring their agent function body, so that they recieve the correct FLAME_DEVICE_API, although this might be avoidable by making that a generic template instantiation. It will simply remove benefits of intellisense.

This version doesn't yet propose how to augment the message accessing based on message type, or proper initialise the message handlers.