Install nvrtc on windows, and set CUDA_PATH and PATH correctly.
The packages installed by the silent installer do not set CUDA_PATH or PATH, so this must be done manually.

CMAKE <= 3.16 does not set CUDA variables correctly for visual studio generators, so the CUDA install must be on the path for the relevant include and lib directories to be found.