CUDAAgent/CUDAAgentStateList changes for submodels
I've gone through all CUDAAgent methods and applied changes to handle a dependent mapped subagent.

It currently doesn't support agent transitions from mapped to unmapped states and vice versa.

I haven't tested this either, but it builds and lints.

I'm trying to workout a way to restructure all of this to be alot clearer too.