Change -DSMS to -DCUDA_ARCH and make cmake more dynamic
This uses nvcc --help to validate if a provide CUDA_ARCH is valid for that version of nvcc.
Default values are stored as a list, which is also validated against the current CUDA version.

Warnings are issued for each user-provided CUDA_ARCH which is not appropriate for the current nvcc.

Also changes the MIN_ARCH compile time macro to MIN_CUDA_ARCH for concistency, and slightly improves the use of it.

Closes #278