HostAgentInstance::sort()
Adds the ability to sort agents via a one or two integer or floating point agent variables, during host functions.

This uses cub device radix sort, which provides no guarantee of a stable sort.

Includes 10 tests, these test both of the new methods with integer and float variables in a full range of orders.
The tests ensure that the key variable becomes sorted, and that other agent variables are reordered with the key variable.

Varadic support was not feasible to include without a hard to understand usage interface, similarly it is unlikely people will require sorting over more than 2 agent variables.

Closes #52