Bucket Messaging format (and associated tests).
Messages are assigned a key which is an integer within a user-defined range.
Multiple messages may have the same key.
This is very similar to how spatial messages work behind the scenes.
Iterator allows access to messages within a specific bin, or a range of contiguous bins.
Note, range iterator is inclusive-exclusive range, whereas description uses inclusive-inclusive range. This might want changing for parity in future.

Closes #82