Improve SWIG exceptions
Python now writes out their type/message by default when the exception is unhandled.

Closes #441

It still outputs pyflamegpu.pyflamegpu.FGPURuntimeException: at the start, as this is the pythone exception class we use.

Converting that to the relevant exception type would require something hacky (or unstable), as SWIG doesn't provide a dynamic way to convert any object to it's PyObject.

Similarly, I couldn't find a way to override that by setting _repr_ or similar.