Adds automatic ID variable to all agents.
This is managed by FLAMEGPU and can be accessed with getID() in all places where getVariable is available.
IDs are assigned when a simulation begins, but can be exported and reimported to/from AgentVector/file.
Users cannot change or disable agent IDs.
id_t is the type of ID, is currently a typedef to unsigned int, but can be changed to uint64_t if desired.
If an id_t type agent/message/env variable is required in Python, newVariableID() or similar can be used.

Closes #199