New experimental CMake flag USE_GLM
Enabling this flag adds the main GLM header to the include stack and allows GLM types to be used to access array variables.
This feature will remain experimental for the time being, whilst we consider the implications.
e.g. __ldg() can't be used for all glm type (e.g. uvec3), so have disabled it with USE_GLM currently.
This commit also include RTC and Non-RTC tests of the functionality.