Add CI Workflow to create draft releases from pushed tags
Also prevents other workflows running on tag pushes.
Adds a CMake variable to specify an exact python version to search for. This is required for windows github actions, but otherwise shouldn't be neccesary
Uses CUDA 11.0 and SM35+ for google collab support
Uses GCC 8 for CUDA 11.0, to avoid a compiler segfault on GHA during building of the test suite.
The Test suite is only built for some entries in the build matrix