Update: RTC Compilation now injects the appropriate #line pragma.
This differs if EXPORT_RTC_SOURCES is enabled, so that line numbers are correct relative to the exported file (which contains injected includes too).

Otherwise, line number is at the start of the user defined RTC agent function string, however if the line begins with a linebreak, that opening linebreak is trimmed.

Closes #608