Specialise python distribution name and embed cuda version in python local build number
pyflamegpu == console
pyflamegpu_vis == with visualisation

Local version number +cuda110 indicates this was build with cuda 11.0 (and therfore depends onit).
This allows wheels with different cuda version to be distinguished
and IF we provide a html file contianig links to all the hosted wheels, we can act as a downstream provider of ourselves.
I.e. python3 -m pip install -f <some address/flamegpu.html> --pre pyflamegpu-v2.0.0-alpha.1+cuda110 would install the cuda 11.0 variant, while +cuda112 would refer to the 11.2 variant.
When uploading to pypi, this would need to be omitted, depending on which version we choose as default. We could handle this via CMake explciticly, or set the new CMake option to OFF for that build.