Add swig wrapper dependencies on flamegpu header files
Generation of the swig wrapper file was not dependent on flamegpu header files, so changes to header files might not cause a regeneration of the wrapper.

CMake 3.22+ would be required for automatic dependency generation via Swigs -M option, but we want to maintain 3.18 minimum.

Fortunately as we have an explciit list of our header files anyway its easy enough to explicitly list these via teh DEPENDS swig source property

Closes #979