Telemetry (#987)
Adds telemetry to the software via the TelemetryDeck web service. This is an important feature for evidencing usage of the software. Without evidence of impact the software will not be able to apply for funding to support its continued development. After consultation with the RSE community the software will be opt-in to enable sharing of usage data. There are various way in which telemetry can be enabled and various environment variables which can be set. 

Makes system calls out to curl with a short timeout, silently fails if not available or it times out. 

Telemetry is opt-out, via an environment variable FLAMEGPU_SHARE_USAGE_STATISTICS=OFF, or by the same value at CMake configuration time.  If disabled at configuraiton time, cmake will emit an message encouraging users to enable it. 

If not enabled at runtime, users will be encouraged to  enable it (once and only once), unless  
FLAMEGPU_TELEMETRY_SILENCE_NOTIFICATION is defined in the environment or at cmake configuraiton time. 

FLAMEGPU develepors can also enable FLAMEGPU_TELEMETRY_TEST_MODE to stop thier telemetry information from being combined with the general users of telemetry.

A random  36 character string is created during CMake Configuration and packaged into the static library, which is used to appropximate unique users. All users of a distributed wheel will be identified as teh same user using this approach, but it avoids encoding system/user information into the telemetry request.

Co-authored-by: Peter Heywood <peethwd@gmail.com>