CI: Add FLAMEGPU_SEATBELTS=OFF CI to the ubuntu and windows regular workflows
Using a number of exclude statemements, per OS the CI matrix should be:

+ 3 release console builds
+ 2 release vis builds
+ 1 beltsoff build

The matrix YAML is not techincally supported by Github actions according to the web editor
We should probably find a compliant way to do this, but it works so a problem for the future.

The individual example uses the same (relevant) options as the main configuration.