WIP: Implement approx equal floating point comparison method
Uses scaled machine epsilon to compare two floats to detertmine if spatial env width is a multiple of the radius.

Only in a test suite so far, with some tests showing it seems to work for previously bad cases.

Todo: Move into the spatial messaging classes, compute on the host, copy if wrapped is available to device, query at runtime in seatbelts mode.