CI: Temporary workaround for manylinux2014 containers glibc error
Github switched from Node 16 to Node 20 for Github actions runners on 2024-06-03.
This is incompatble with older linux distributions, including centos7.

This sets an env var which will alllow the use of node 16, but this will stop working at some point when github fully remove node 16.

This is separate but related to #1224