Fix: Improve wrapped spatial messaging bounds/radius compatibility checking
Improves checking that the communication radius is a factor of the envionment dimensions when using spaical communication with seatbelts enabled

This was preivously not always correct due to floating point use for some common cases, including 1 / 0.05f.

This is now checked once on the host, and stored in the messageing MetaData so that if the wrapped communicator is used on device a runtime exception can be raised.

A new internal (detail) header is creatd with a function that is common to 2D and 3D implementations.

Closes #1177